backend/requirements.txt
blinker==1.8.2
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
filelock==3.13.1
Flask==3.0.3
Flask-Cors==4.0.1
fsspec==2024.2.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
mpmath==1.3.0
mutagen==1.47.0
networkx==3.2.1
numpy==2.0.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.20.5
nvidia-nvtx-cu11==11.8.86
opencv-python==4.10.0.84
pillow==10.2.0
pycryptodomex==3.20.0
pytube==15.0.0
pytubefix==6.10.2
requests==2.32.3
setuptools==70.0.0
sympy==1.12
torch==2.4.0+cu118
torchaudio==2.4.0+cu118
torchvision==0.19.0+cu118
tqdm==4.66.5
triton==3.0.0
typing_extensions==4.9.0
urllib3==2.2.2
websockets==12.0
Werkzeug==3.0.3
